-r doesn't work, reverting to -i
Tested on ubuntu 20.04, anaconda python 3.7.1, TShark (Wireshark) 3.2.3 (Git v3.2.3 packaged as 3.2.3-1).
https://github.com/KimiNewt/pyshark/issues/477
https://github.com/KimiNewt/pyshark/issues/299
with LiveRingCapture and -r it just crashed, as in these issues. With LiveCapture it would drop packets. With LiveRingCapture and -i it works without dropping packets. I did not run for a long period of time yet to see how long it runs without drops. My packet rate is relatively low.